server/requirements.txt
fastapi==0.99.1
uvicorn
openai
python-dotenv
motor
pydantic
python-jose[cryptography]
beanie
passlib[bcrypt]
pyjwt
bcrypt
pytest-asyncio
httpx
google-auth
requests
boto3
black
isort